# api/openapi/vmafx-server-v1.yaml
openapi: "3.0.3"
info:
  title: VMAFX Scoring API
  description: |
    REST API for VMAFX perceptual video quality scoring.

    The HTTP transport delegates internally to the same libvmaf scorer that
    backs the gRPC service defined in `proto/vmafx.proto`.  All REST endpoints
    are additive: existing gRPC clients are unaffected.

    **Versioning**: this spec tracks the `v1` resource prefix.  Breaking
    changes increment the major version and are announced via ADR.

    See [docs/server/rest.md](../../docs/server/rest.md) for operator guidance.
  version: "1.0.0"
  license:
    name: EUPL-1.2
    url: https://github.com/VMAFx/vmafx/blob/master/LICENSES/EUPL-1.2.txt
  contact:
    name: VMAFX fork
    url: https://github.com/VMAFx/vmafx

servers:
  - url: http://localhost:8080
    description: Local development instance

tags:
  - name: scoring
    description: VMAF perceptual quality scoring operations
  - name: health
    description: Liveness and readiness probes

paths:
  /v1/score:
    post:
      operationId: scoreVideoPair
      summary: Score a reference/distorted video pair
      description: |
        Computes the aggregate VMAF score and per-feature breakdown for a
        reference/distorted video pair.  Both `reference` and `distorted`
        must be paths resolvable by the server process (absolute paths or
        paths relative to the server working directory).

        The optional `model` field selects the VMAF model JSON filename
        (without the `.json` extension) from `VMAFX_MODEL_DIR`.  When
        omitted the server defaults to `vmaf_v1.0.16_3d0h`, the library
        default (`VMAF_DEFAULT_MODEL_VERSION`).

        The optional `options` object carries the scoring options
        (`ScoreOptions`, generated from the API definition): raw `.yuv`
        geometry, backend and device, threads, subsample, precision,
        features, tiny model and more. Scores are lossless (`precision`
        `max`) unless `options.precision` asks otherwise, and every
        response carries `provenance`. The same request gives the same
        score through the `vmaf` CLI and the C API (docs/server/api-contract.md).
      tags: [scoring]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreRequest"
            examples:
              netflix_golden:
                summary: Netflix golden pair
                value:
                  reference: "/data/src01_hrc00_576x324.yuv"
                  distorted: "/data/src01_hrc01_576x324.yuv"
                  model: "vmaf_v0.6.1"
                  options:
                    width: 576
                    height: 324
                    pixel_format: "420"
                    bitdepth: 8
      responses:
        "200":
          description: Scoring completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreResponse"
              examples:
                netflix_golden:
                  summary: Netflix golden result
                  value:
                    score: 76.6683
                    features:
                      adm2: 0.9876
                      motion2: 2.3456
                      vif_scale0: 0.8912
        "400":
          description: Missing or invalid request fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_fields:
                  value:
                    error: "reference and distorted are required"
        "500":
          description: Internal scoring error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/health:
    get:
      operationId: getHealth
      summary: Liveness probe
      description: |
        Always returns `200 OK` while the server process is live.
        Kubernetes liveness probes should target this endpoint.

        Note: `/healthz` (without the `/v1/` prefix) is an alias kept for
        backwards compatibility with Kubernetes conventions.
      tags: [health]
      responses:
        "200":
          description: Server is live
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                ok:
                  value:
                    status: "ok"

  /v1/ready:
    get:
      operationId: getReady
      summary: Readiness probe
      description: |
        Returns `200 OK` once the scorer is initialised and ready to accept
        scoring requests.  Returns `503 Service Unavailable` otherwise.

        Note: `/readyz` (without the `/v1/` prefix) is an alias kept for
        backwards compatibility with Kubernetes conventions.
      tags: [health]
      responses:
        "200":
          description: Server is ready to accept scoring requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"
              examples:
                ready:
                  value:
                    status: "ready"
        "503":
          description: Scorer not yet initialised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"
              examples:
                not_ready:
                  value:
                    status: "not ready"
                    reason: "scorer not initialised"

components:
  schemas:
    # BEGIN GENERATED: vmafx-api scoring schemas (scripts/codegen/vmafx-api.py)
    ScoreOptions:
      type: "object"
      description: "Scoring options."
      properties:
        width:
          type: "integer"
          minimum: 1
          format: "int32"
          description: "Width of raw .yuv input in pixels."
        height:
          type: "integer"
          minimum: 1
          format: "int32"
          description: "Height of raw .yuv input in pixels."
        pixel_format:
          type: "string"
          enum:
            - "420"
            - "422"
            - "444"
          description: "Chroma subsampling of raw .yuv input. One of: 420, 422, 444."
        bitdepth:
          type: "integer"
          enum:
            - 8
            - 10
            - 12
            - 16
          format: "int32"
          description: "Bits per sample of raw .yuv input. One of: 8, 10, 12, 16."
        disable_clip:
          type: "boolean"
          description: "Do not clip the model score to [0, 100] (the model's disable_clip). Unset: false."
        enable_transform:
          type: "boolean"
          description: "Apply the model's score transform (the model's enable_transform). Unset: false."
        backend:
          type: "string"
          enum:
            - "auto"
            - "cpu"
            - "cuda"
            - "sycl"
            - "hip"
            - "metal"
          description: "Backend: auto uses the available ones; any other value runs that backend alone and fails when it is not available. One of: auto, cpu, cuda, sycl, hip, metal. Unset: auto."
        feature:
          type: "array"
          items:
            type: "string"
          description: "Additional feature extractor, name[=key=value:...] (for example psnr or cambi=full_ref=true); several may be given (the filter separates them with |). Mutually exclusive with the CTC presets."
        aom_ctc:
          type: "string"
          enum:
            - "v1.0"
            - "v2.0"
            - "v3.0"
            - "v4.0"
            - "v5.0"
            - "v6.0"
            - "v7.0"
          description: "AOM common test conditions preset: a fixed model and feature set. One of: v1.0, v2.0, v3.0, v4.0, v5.0, v6.0, v7.0."
        nflx_ctc:
          type: "string"
          enum:
            - "v1.0"
          description: "Netflix common test conditions preset: a fixed model and feature set. One of: v1.0."
        tiny_model:
          type: "string"
          description: "Tiny ONNX model to load alongside the classic models."
        tiny_device:
          type: "string"
          enum:
            - "auto"
            - "cpu"
            - "cuda"
            - "openvino"
            - "openvino-npu"
            - "openvino-cpu"
            - "openvino-gpu"
            - "coreml"
            - "coreml-ane"
            - "coreml-gpu"
            - "coreml-cpu"
            - "rocm"
          description: "ONNX Runtime execution provider of the tiny model. One of: auto, cpu, cuda, openvino, openvino-npu, openvino-cpu, openvino-gpu, coreml, coreml-ane, coreml-gpu, coreml-cpu, rocm. Unset: auto."
        tiny_threads:
          type: "integer"
          minimum: 0
          format: "int32"
          description: "Intra-op threads of the CPU execution provider (0: the runtime's default)."
        tiny_fp16:
          type: "boolean"
          description: "Request fp16 input and output where the execution provider supports it. Unset: false."
        tiny_model_verify:
          type: "boolean"
          description: "Require a Sigstore bundle verification of the tiny model (cosign verify-blob) before it loads; a missing bundle, a missing cosign or a failed verification refuses the model. Unset: false."
        tiny_codec:
          type: "string"
          description: "Encoder of the distorted clip, required by codec-aware tiny models (fr_regressor_v2/v3), which refuse to score without it. Must be in the model sidecar's encoder_vocab; the ffprobe names h264, hevc, av1, vp9 and vvc are accepted."
        tiny_preset:
          type: "string"
          description: "Encoder preset (medium, slow, p4, 5, ...), read as the encoder defines it. Unset: ordinal 5 (medium). A model trained with one preset (fr_regressor_v3) ignores it and warns."
        tiny_crf:
          type: "integer"
          minimum: 0
          maximum: 63
          format: "int32"
          description: "CRF or QP used for the encode, normalised as the model sidecar declares. Required with the codec and preset."
        tiny_resize:
          type: "string"
          enum:
            - "bilinear"
            - "nearest"
            - "bicubic"
            - "disabled"
          description: "Resize filter for NCHW tiny models whose input size differs from the frame; disabled refuses the mismatch (-ERANGE). The three filters give scores about 2% apart: record the filter with the model. One of: bilinear, nearest, bicubic, disabled. Unset: disabled."
        no_reference:
          type: "boolean"
          description: "No-reference mode; needs a no-reference tiny model. The reference becomes a formality: only the distorted picture is scored. Unset: false."
        threads:
          type: "integer"
          minimum: 0
          format: "int32"
          description: "Worker threads of the feature extractors, capped to the hardware threads (0: score in the calling thread)."
        frame_cnt:
          type: "integer"
          minimum: 1
          format: "int32"
          description: "Score at most this many frames."
        frame_skip_ref:
          type: "integer"
          minimum: 0
          format: "int32"
          description: "Skip this many frames at the start of the reference."
        frame_skip_dist:
          type: "integer"
          minimum: 0
          format: "int32"
          description: "Skip this many frames at the start of the distorted video."
        no_prediction:
          type: "boolean"
          description: "Extract features only; no model score. Unset: false."
        cpumask:
          type: "integer"
          minimum: 0
          format: "int32"
          description: "Bitmask of CPU instruction sets the extractors must not use."
        gpumask:
          type: "integer"
          minimum: 0
          format: "int32"
          description: "Bitmask of GPU operations the extractors must not use."
        device:
          type: "string"
          description: "GPU of the selected backend: auto, or a device index. Needs a GPU backend that selects devices by index (sycl, hip, metal). Unset: auto."
        subsample:
          type: "integer"
          minimum: 1
          format: "int32"
          description: "Score every n-th frame (1: every frame). Unset: 1."
        precision:
          type: "string"
          enum:
            - "legacy"
            - "max"
            - "full"
            - "1"
            - "2"
            - "3"
            - "4"
            - "5"
            - "6"
            - "7"
            - "8"
            - "9"
            - "10"
            - "11"
            - "12"
            - "13"
            - "14"
            - "15"
            - "16"
            - "17"
          description: "Score precision: N (1 to 17) writes %.<N>g; max or full write %.17g (round-trip lossless); legacy writes %.6f (Netflix-compatible). The scoring server returns lossless scores unless asked otherwise. One of: legacy, max, full, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17. Unset: max."
        view_distance:
          type: "number"
          minimum: 0.75
          maximum: 24.0
          format: "double"
          description: "Viewing distance in display heights (ADM adm_norm_view_dist). Unset: the model's value."
        display_height:
          type: "integer"
          minimum: 1
          format: "int32"
          description: "Height of the reference display in pixels (ADM adm_ref_display_height). Unset: the model's value."
        target_width:
          type: "integer"
          minimum: 0
          format: "int32"
          description: "Width of the target display the distorted video is scaled to (0: no scaling). Unset: 0. Reserved: device-targeted scoring lands in RC5; only the default is accepted."
        target_height:
          type: "integer"
          minimum: 0
          format: "int32"
          description: "Height of the target display the distorted video is scaled to (0: no scaling). Unset: 0. Reserved: device-targeted scoring lands in RC5; only the default is accepted."
        target_scaling:
          type: "string"
          enum:
            - "none"
            - "bilinear"
            - "bicubic"
            - "lanczos"
          description: "Scaling filter towards the target display. One of: none, bilinear, bicubic, lanczos. Unset: none. Reserved: device-targeted scoring lands in RC5; only the default is accepted."
    Provenance:
      type: "object"
      description: "How the scores of a context are made (#2142). Strings live as long as the context. (from VmafxProvenance)."
      properties:
        abi_major:
          type: "integer"
          format: "int32"
          description: "ABI major version of the library that ran."
        abi_minor:
          type: "integer"
          format: "int32"
          description: "ABI minor version."
        abi_patch:
          type: "integer"
          format: "int32"
          description: "ABI patch version."
        active_backend:
          type: "string"
          description: "Backend imported into the context. Value name of VmafxBackend without the VMAFX_BACKEND_ prefix, lower case."
        n_extractors:
          type: "integer"
          format: "int32"
          description: "Registered feature extractors; see vmafx_context_extractor_info()."
        version:
          type: "string"
          description: "Build version (git describe)."
    # END GENERATED: vmafx-api scoring schemas

    ScoreRequest:
      type: object
      required:
        - reference
        - distorted
      properties:
        reference:
          type: string
          description: |
            Path to the reference (pristine) video file.  YUV 4:2:0, Y4M,
            and common container formats supported by the backing vmaf binary.
          example: "/data/reference.yuv"
        distorted:
          type: string
          description: |
            Path to the distorted (compressed) video file.  Must be the same
            resolution and frame count as the reference.
          example: "/data/distorted.yuv"
        model:
          type: string
          description: |
            VMAF model name to use.  Must match a `.json` filename (without
            extension) in the server's `VMAFX_MODEL_DIR`.  Defaults to
            `vmaf_v1.0.16_3d0h`, the library default, when omitted.
          example: "vmaf_v0.6.1"
        options:
          $ref: "#/components/schemas/ScoreOptions"

    ScoreResponse:
      type: object
      required: [score, features]
      properties:
        score:
          type: number
          format: double
          description: Aggregate VMAF score (0–100 typical range).
          example: 76.6683
        features:
          type: object
          description: |
            Per-feature scores keyed by feature name.  The set of keys depends
            on the selected model; common keys include `adm2`, `motion2`,
            `vif_scale0`, `vif_scale1`, `vif_scale2`, `vif_scale3`.
          additionalProperties:
            type: number
            format: double
          example:
            adm2: 0.9876
            motion2: 2.3456
            vif_scale0: 0.8912
        provenance:
          $ref: "#/components/schemas/ScoreProvenance"

    FeatureBackend:
      type: object
      description: One entry of the backend receipt.
      properties:
        extractor:
          type: string
          description: Extractor name (e.g. `adm`, `motion`).
        backend:
          type: string
          description: Backend it ran on (`cpu`, `cuda`, `sycl`, `hip`, `metal`).

    ScoreProvenance:
      type: object
      description: |
        How a score was made (#2142, #2155): the library record, the model the
        server loaded and the backend receipt. Every scoring response carries it.
      properties:
        library:
          $ref: "#/components/schemas/Provenance"
        model:
          type: string
          description: Model the server resolved the request's model name to.
        model_sha256:
          type: string
          description: SHA-256 (lower-case hex) of the model file the server loaded.
        backend_used:
          type: string
          description: Backend the extractors ran on (`cpu` when all ran on the CPU).
        feature_backends:
          type: array
          items:
            $ref: "#/components/schemas/FeatureBackend"
        precision:
          type: string
          description: Precision the scores were computed with (`max` is lossless).

    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
          description: Always "ok" for a live server.
          example: "ok"

    ReadyResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ready, "not ready"]
          description: >
            "ready" when the scorer is initialised; "not ready" otherwise.
          example: "ready"
        reason:
          type: string
          description: Human-readable explanation when not ready.
          example: "scorer not initialised"

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "reference and distorted are required"